Extract CannonSelector + Refactor + Hotfix /cannons claim
