+++ src/main/java/at/pavlov/cannons/cannon/CannonManager.java
@@ -21,6 +21,7 @@ import org.bukkit.OfflinePlayer;
 import org.bukkit.World;
 import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Player;
+import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.util.Vector;
 
 import java.text.DecimalFormat;
@@ -32,46 +33,47 @@ import java.util.concurrent.ConcurrentHashMap;
 
 
 public class CannonManager {
-	private static final ConcurrentHashMap<UUID, Cannon> cannonList = new ConcurrentHashMap<>();
+    private static final ConcurrentHashMap<UUID, Cannon> cannonList = new ConcurrentHashMap<>();
     private static final ConcurrentHashMap<String, UUID> cannonNameMap = new ConcurrentHashMap<>();
 
 
     private final Cannons plugin;
-	private final UserMessages userMessages;
-	private final Config config;
+    private final UserMessages userMessages;
+    private final Config config;
 
 
+    public CannonManager(Cannons cannons, UserMessages userMessages, Config config) {
+        this.userMessages = userMessages;
+        this.config = config;
+        this.plugin = cannons;
+    }
 
-	public CannonManager(Cannons cannons, UserMessages userMessages, Config config) {
-		this.userMessages = userMessages;
-		this.config = config;
-		this.plugin = cannons;
-	}
-
-	/**
-	 * removes a cannons from the list that are not valid
+    /**
+     * removes a cannons from the list that are not valid
+     *
      * @param cause the reason why the cannon is removed
-	 */
-	private void removeInvalidCannons(BreakCause cause) {
-		Iterator<Cannon> iter = cannonList.values().iterator();
+     */
+    private void removeInvalidCannons(BreakCause cause) {
+        Iterator<Cannon> iter = cannonList.values().iterator();
 
-        while(iter.hasNext()) {
+        while (iter.hasNext()) {
             Cannon next = iter.next();
-			if (!next.isValid()) {
-				removeCannon(next, false, false, cause, false, false);
+            if (!next.isValid()) {
+                removeCannon(next, false, false, cause, false, false);
                 iter.remove();
-			}
-		}
-	}
+            }
+        }
+    }
 
     /**
      * deconstructs a cannon without the risk of explosion
+     *
      * @param cannon cannon to remove
      */
     public void dismantleCannon(Cannon cannon, Player player) {
         if (cannon == null)
             return;
-        if (player==null){
+        if (player == null) {
             removeCannon(cannon, true, false, BreakCause.Dismantling);
             return;
         }
@@ -83,7 +85,7 @@ public class CannonManager {
 
         if (player.hasPermission(cannon.getCannonDesign().getPermissionDismantle())) {
             //only the owner of the cannon can dismantle a cannon
-            if (cannon.getOwner()!=null && cannon.getOwner().equals(player.getUniqueId()))
+            if (cannon.getOwner() != null && cannon.getOwner().equals(player.getUniqueId()))
                 removeCannon(cannon, true, false, BreakCause.Dismantling);
             else
                 userMessages.sendMessage(MessageEnum.ErrorDismantlingNotOwner, player, cannon);
@@ -93,35 +95,38 @@ public class CannonManager {
         userMessages.sendMessage(MessageEnum.PermissionErrorDismantle, player, cannon);
     }
 
-	/**
-	 * removes a cannon from the list
-	 * @param loc location of the cannon
+    /**
+     * removes a cannon from the list
+     *
+     * @param loc         location of the cannon
      * @param breakCannon all cannon blocks will drop
-     * @param canExplode if the cannon can explode when loaded with gunpowder
-     * @param cause the reason way the cannon was broken
-	 */
-	public void removeCannon(Location loc, boolean breakCannon, boolean canExplode, BreakCause cause) {
-		Cannon cannon = getCannon(loc, null);
-		removeCannon(cannon, breakCannon, canExplode, cause);
-	}
-
-	/**
-	 * removes a cannon from the list
-	 * @param cannon cannon to remove
-     * @param breakCannon all cannon blocks will drop
-     * @param canExplode if the cannon can explode when loaded with gunpowder
-     * @param cause the reason way the cannon was broken
-	 */
-	public void removeCannon(Cannon cannon, boolean breakCannon, boolean canExplode, BreakCause cause) {
-        removeCannon(cannon, breakCannon, canExplode, cause, true, true);
-	}
+     * @param canExplode  if the cannon can explode when loaded with gunpowder
+     * @param cause       the reason way the cannon was broken
+     */
+    public void removeCannon(Location loc, boolean breakCannon, boolean canExplode, BreakCause cause) {
+        Cannon cannon = getCannon(loc, null);
+        removeCannon(cannon, breakCannon, canExplode, cause);
+    }
 
     /**
      * removes a cannon from the list
-     * @param uid UID of the cannon
+     *
+     * @param cannon      cannon to remove
      * @param breakCannon all cannon blocks will drop
-     * @param canExplode if the cannon can explode when loaded with gunpowder
-     * @param cause the reason way the cannon was broken
+     * @param canExplode  if the cannon can explode when loaded with gunpowder
+     * @param cause       the reason way the cannon was broken
+     */
+    public void removeCannon(Cannon cannon, boolean breakCannon, boolean canExplode, BreakCause cause) {
+        removeCannon(cannon, breakCannon, canExplode, cause, true, true);
+    }
+
+    /**
+     * removes a cannon from the list
+     *
+     * @param uid         UID of the cannon
+     * @param breakCannon all cannon blocks will drop
+     * @param canExplode  if the cannon can explode when loaded with gunpowder
+     * @param cause       the reason way the cannon was broken
      */
     public void removeCannon(UUID uid, boolean breakCannon, boolean canExplode, BreakCause cause) {
         removeCannon(uid, breakCannon, canExplode, cause, true);
@@ -129,10 +134,11 @@ public class CannonManager {
 
     /**
      * removes a cannon from the list
-     * @param uid UID of the cannon
+     *
+     * @param uid         UID of the cannon
      * @param breakCannon all cannon blocks will drop
-     * @param canExplode if the cannon can explode when loaded with gunpowder
-     * @param cause the reason way the cannon was broken
+     * @param canExplode  if the cannon can explode when loaded with gunpowder
+     * @param cause       the reason way the cannon was broken
      * @param removeEntry should the cannon be removed from the list
      */
     public void removeCannon(UUID uid, boolean breakCannon, boolean canExplode, BreakCause cause, boolean removeEntry) {
@@ -143,10 +149,11 @@ public class CannonManager {
 
     /**
      * removes a cannon from the list
-     * @param cannon cannon to remove
-     * @param breakCannon all cannon blocks will drop
-     * @param canExplode if the cannon can explode when loaded with gunpowder
-     * @param cause the reason way the cannon was broken
+     *
+     * @param cannon        cannon to remove
+     * @param breakCannon   all cannon blocks will drop
+     * @param canExplode    if the cannon can explode when loaded with gunpowder
+     * @param cause         the reason way the cannon was broken
      * @param ignoreInvalid if true invalid cannons will be skipped and not removed
      */
     public void removeCannon(Cannon cannon, boolean breakCannon, boolean canExplode, BreakCause cause, boolean removeEntry, boolean ignoreInvalid) {
@@ -158,9 +165,8 @@ public class CannonManager {
         if (cause == BreakCause.Dismantling || cause == BreakCause.Other) {
             plugin.logDebug("Dismantling," + cannon.getCannonDesign().getSoundDismantle().toString());
             CannonsUtil.playSound(cannon.getRandomBarrelBlock(), cannon.getCannonDesign().getSoundDismantle());
-            delay = (long) (cannon.getCannonDesign().getDismantlingDelay()*20.0);
-        }
-        else
+            delay = (long) (cannon.getCannonDesign().getDismantlingDelay() * 20.0);
+        } else
             CannonsUtil.playSound(cannon.getRandomBarrelBlock(), cannon.getCannonDesign().getSoundDestroy());
 
         //delay the remove task, so it fits to the sound
@@ -213,47 +219,49 @@ public class CannonManager {
     }
 
 
-	/**
-	 * Checks if the name of a cannon is unique
-	 * @param name name of the cannon
-	 * @return true if the name is unique
-	 */
-	private static boolean isCannonNameUnique(String name) {
+    /**
+     * Checks if the name of a cannon is unique
+     *
+     * @param name name of the cannon
+     * @return true if the name is unique
+     */
+    private static boolean isCannonNameUnique(String name) {
         if (name == null)
             return false;
 
-		// try to find this in the map
+        // try to find this in the map
         //there is no such cannon name
         //there is such a cannon name
         return cannonNameMap.get(name) == null;
-	}
+    }
 
-	/**
-	 * generates a new unique cannon name
-	 * @return name string for the new cannon
-	 */
-	private String newCannonName(Cannon cannon) {
-		//check if this cannon has a owner
-		if (cannon.getOwner() == null)
+    /**
+     * generates a new unique cannon name
+     *
+     * @return name string for the new cannon
+     */
+    private String newCannonName(Cannon cannon) {
+        //check if this cannon has a owner
+        if (cannon.getOwner() == null)
             return "missing Owner";
-			
-		String name;
-		CannonDesign design = cannon.getCannonDesign();
-		if (design != null)
-			name = design.getDesignName();
-		else
-			name = "cannon";
-	
 
-		for (int i = 1; i < Integer.MAX_VALUE; i++) {
-			String cannonName = name + " " + i;
+        String name;
+        CannonDesign design = cannon.getCannonDesign();
+        if (design != null)
+            name = design.getDesignName();
+        else
+            name = "cannon";
 
-			if (isCannonNameUnique(cannonName)) {
-				return cannonName;
-			}
-		}
-		return "no unique name";
-	}
+
+        for (int i = 1; i < Integer.MAX_VALUE; i++) {
+            String cannonName = name + " " + i;
+
+            if (isCannonNameUnique(cannonName)) {
+                return cannonName;
+            }
+        }
+        return "no unique name";
+    }
 
     public MessageEnum renameCannon(Player player, Cannon cannon, String newCannonName) {
         Validate.notNull(player, "player must not be null");
@@ -281,17 +289,18 @@ public class CannonManager {
 
     }
 
-	/**
-	 * adds a new cannon to the list of cannons
-	 * @param cannon create this cannon
+    /**
+     * adds a new cannon to the list of cannons
+     *
+     * @param cannon         create this cannon
      * @param saveToDatabase if the cannon will be saved to the database after loading
-	 */
-	public void createCannon(Cannon cannon, boolean saveToDatabase) {
+     */
+    public void createCannon(Cannon cannon, boolean saveToDatabase) {
         //the owner can't be null
-		if (cannon.getOwner() == null) {
-			plugin.logInfo("can't save a cannon when the owner is null");
-			return;
-		}
+        if (cannon.getOwner() == null) {
+            plugin.logInfo("can't save a cannon when the owner is null");
+            return;
+        }
 
         //ignore paid if there is no economy
         if (plugin.getEconomy() == null || cannon.getCannonDesign().getEconomyBuildingCost() <= 0)
@@ -300,13 +309,13 @@ public class CannonManager {
         //add owner to whitelist for sentry
         if (cannon.getCannonDesign().isSentry())
             cannon.addWhitelistPlayer(cannon.getOwner());
-		
-		//if the cannonName is empty make a new one
-		if (cannon.getCannonName() ==  null || cannon.getCannonName().isEmpty())
-			cannon.setCannonName(newCannonName(cannon));
 
-		// add cannon to the list
-		cannonList.put(cannon.getUID(), cannon);
+        //if the cannonName is empty make a new one
+        if (cannon.getCannonName() == null || cannon.getCannonName().isEmpty())
+            cannon.setCannonName(newCannonName(cannon));
+
+        // add cannon to the list
+        cannonList.put(cannon.getUID(), cannon);
         //add cannon name to the list
         cannonNameMap.put(cannon.getCannonName(), cannon.getUID());
         if (cannon.getCannonDesign().isSentry())
@@ -323,11 +332,12 @@ public class CannonManager {
 
         LoadWhitelistTask loadWhitelistTask = new LoadWhitelistTask(cannon.getUID());
         loadWhitelistTask.runTaskAsynchronously(plugin);
-	}
+    }
 
     /**
      * returns all known cannons in a sphere around the given location
-     * @param center - center of the box
+     *
+     * @param center       - center of the box
      * @param sphereRadius - radius of the sphere in blocks
      * @return - list of all cannons in this sphere
      */
@@ -349,7 +359,8 @@ public class CannonManager {
 
     /**
      * returns all known cannons in a box around the given location
-     * @param center - center of the box
+     *
+     * @param center  - center of the box
      * @param lengthX - box length in X
      * @param lengthY - box length in Y
      * @param lengthZ - box length in Z
@@ -371,19 +382,33 @@ public class CannonManager {
         return newCannonList;
     }
 
+    //please nobody call this
     public void claimCannonsInBox(Location center, UUID owner) {
-        int halflength = 60;
-        for (int x = halflength; x >= -halflength; x--) {
-            for (int y = halflength; y >= -halflength; y--) {
-                for (int z = halflength; z >= -halflength; z--) {
-                    getCannon(center.clone().add(x, y, z), owner);
+        try {
+            new BukkitRunnable() {
+                @Override
+                public void run() {
+                    int halflength = 20;
+                    for (int x = halflength; x >= -halflength; x--) {
+                        for (int y = halflength; y >= -halflength; y--) {
+                            for (int z = halflength; z >= -halflength; z--) {
+                                getCannon(center.clone().add(x, y, z), owner);
+                            }
+                        }
+                    }
                 }
-            }
+            }.runTaskAsynchronously(Cannons.getPlugin());
+        } catch (Exception e) {
+            System.err.println(e.getClass());
+            e.printStackTrace();
+
         }
+
     }
 
     /**
      * returns all cannons for a list of locations
+     *
      * @param locations - a list of location to search for cannons
      * @return - list of all cannons in this sphere
      */
@@ -400,9 +425,10 @@ public class CannonManager {
 
     /**
      * returns all cannons for a list of locations - this will update all locations
+     *
      * @param locations - a list of location to search for cannons
-     * @param player - player which operates the cannon
-     * @param silent - no messages will be displayed if silent is true
+     * @param player    - player which operates the cannon
+     * @param silent    - no messages will be displayed if silent is true
      * @return - list of all cannons in this sphere
      */
     public HashSet<Cannon> getCannons(List<Location> locations, UUID player, boolean silent) {
@@ -417,56 +443,59 @@ public class CannonManager {
         return newCannonList;
     }
 
-	/**
-	 * get cannon by cannonName and Owner - used for Signs
-	 * @param cannonName name of the cannon
-	 * @return the cannon with this name
-	 */
-	public static Cannon getCannon(String cannonName) {
-		if (cannonName == null) return null;
+    /**
+     * get cannon by cannonName and Owner - used for Signs
+     *
+     * @param cannonName name of the cannon
+     * @return the cannon with this name
+     */
+    public static Cannon getCannon(String cannonName) {
+        if (cannonName == null) return null;
 
         UUID uid = cannonNameMap.get(cannonName);
         if (uid != null)
-		    return cannonList.get(uid);
+            return cannonList.get(uid);
 
         //cannon not found
         return null;
-	}
+    }
 
-	/**
-	 * Searches the storage if there is already a cannonblock on this location
-	 * and returns the cannon
-	 * @param loc location of one cannon block
-	 * @return the cannon at this location
-	 */
-	private Cannon getCannonFromStorage(Location loc) {
-		for (Cannon cannon : cannonList.values()) {
+    /**
+     * Searches the storage if there is already a cannonblock on this location
+     * and returns the cannon
+     *
+     * @param loc location of one cannon block
+     * @return the cannon at this location
+     */
+    private Cannon getCannonFromStorage(Location loc) {
+        for (Cannon cannon : cannonList.values()) {
             //To make code faster on servers with a lot of cannons we check the distance squared
-            if (loc.toVector().distanceSquared(cannon.getOffset()) <= 1024 && cannon.isCannonBlock(loc.getBlock()))  {
-				return cannon;
-			}
-		}
-		return null;
-	}
+            if (loc.toVector().distanceSquared(cannon.getOffset()) <= 1024 && cannon.isCannonBlock(loc.getBlock())) {
+                return cannon;
+            }
+        }
+        return null;
+    }
 
-	/**
-	 * searches for a cannon and creates a new entry if it does not exist
-	 * @param cannonBlock - one block of the cannon
-	 * @param owner - the owner of the cannon (important for message notification). Can't be null if a new cannon is created
-	 * @return the cannon at this location
-	 */
-	public Cannon getCannon(Location cannonBlock, UUID owner) {
-		return getCannon(cannonBlock, owner, false);
-	}
-	
-	/**
-	 * searches for a cannon and creates a new entry if it does not exist
-	 * 
-	 * @param cannonBlock - one block of the cannon
-	 * @param owner - the owner of the cannon (important for message notification). Can't be null
-	 * @return the cannon at this location
-	 */
-	public Cannon getCannon(Location cannonBlock, UUID owner, boolean silent) {
+    /**
+     * searches for a cannon and creates a new entry if it does not exist
+     *
+     * @param cannonBlock - one block of the cannon
+     * @param owner       - the owner of the cannon (important for message notification). Can't be null if a new cannon is created
+     * @return the cannon at this location
+     */
+    public Cannon getCannon(Location cannonBlock, UUID owner) {
+        return getCannon(cannonBlock, owner, false);
+    }
+
+    /**
+     * searches for a cannon and creates a new entry if it does not exist
+     *
+     * @param cannonBlock - one block of the cannon
+     * @param owner       - the owner of the cannon (important for message notification). Can't be null
+     * @return the cannon at this location
+     */
+    public Cannon getCannon(Location cannonBlock, UUID owner, boolean silent) {
         // is this block material used for a cannon design
         if (cannonBlock.getBlock() == null || !plugin.getDesignStorage().isCannonBlockMaterial(cannonBlock.getBlock().getType()))
             return null;
@@ -492,18 +521,18 @@ public class CannonManager {
             //use the updated object from the storage
             cannon = cannonFromSign;
 
-            plugin.logDebug("Time to find cannon: " + new DecimalFormat("0.00").format((System.nanoTime() - startTime)/1000000.0) + "ms");
+            plugin.logDebug("Time to find cannon: " + new DecimalFormat("0.00").format((System.nanoTime() - startTime) / 1000000.0) + "ms");
             return cannon;
         }
 
         // this cannon has no sign, so look in the database if there is something
-        Cannon storageCannon =  getCannonFromStorage(cannonBlock);
+        Cannon storageCannon = getCannonFromStorage(cannonBlock);
         if (storageCannon != null) {
             //try to find something in the storage
             plugin.logDebug("cannon found in storage");
             cannon = storageCannon;
 
-            plugin.logDebug("Time to find cannon: " + new DecimalFormat("0.00").format((System.nanoTime() - startTime)/1000000.0) + "ms");
+            plugin.logDebug("Time to find cannon: " + new DecimalFormat("0.00").format((System.nanoTime() - startTime) / 1000000.0) + "ms");
             return cannon;
         } //nothing in the storage, so we make a new entry
 
@@ -513,7 +542,7 @@ public class CannonManager {
         Player player = Bukkit.getPlayer(owner);
 
         //can this player can build one more cannon
-        MessageEnum	message = canBuildCannon(cannon, owner);
+        MessageEnum message = canBuildCannon(cannon, owner);
 
         //check the permissions for redstone
         if (message == null || message == MessageEnum.CannonCreated)
@@ -524,9 +553,9 @@ public class CannonManager {
             message = MessageEnum.ErrorMissingSign;
 
         Cannon created = startCannonCreation(cannon, message, player, owner, silent);
-        plugin.logDebug("Time to find cannon: " + new DecimalFormat("0.00").format((System.nanoTime() - startTime)/1000000.0) + "ms");
+        plugin.logDebug("Time to find cannon: " + new DecimalFormat("0.00").format((System.nanoTime() - startTime) / 1000000.0) + "ms");
         return created;
-	}
+    }
 
     private Cannon startCannonCreation(Cannon cannon, MessageEnum message, Player player, UUID owner, boolean silent) {
         plugin.logDebug("CannonBeforeCreateEvent Cannon: " + cannon + "message: " + message + " player: " + player);
@@ -563,6 +592,7 @@ public class CannonManager {
 
     /**
      * returns the cannon from the storage
+     *
      * @param uid UUID of the cannon
      * @return the cannon from the storage
      */
@@ -573,34 +603,35 @@ public class CannonManager {
         return cannonList.get(uid);
     }
 
-	/**
-	 * searches if this block is part of a cannon and create a new one
-	 * @param cannonBlock block of the cannon
-	 * @param owner the player who will be the owner of the cannon if it is a new cannon
-	 * @return cannon if found, else null
-	 */
+    /**
+     * searches if this block is part of a cannon and create a new one
+     *
+     * @param cannonBlock block of the cannon
+     * @param owner       the player who will be the owner of the cannon if it is a new cannon
+     * @return cannon if found, else null
+     */
     private Cannon checkCannon(Location cannonBlock, UUID owner) {
-	    // is this block material used for a cannon design
+        // is this block material used for a cannon design
         if (cannonBlock.getBlock() == null || !plugin.getDesignStorage().isCannonBlockMaterial(cannonBlock.getBlock().getType()))
             return null;
 
-		World world = cannonBlock.getWorld();
+        World world = cannonBlock.getWorld();
 
-		// check all cannon design if this block is part of the design
-		for (CannonDesign cannonDesign : plugin.getDesignStorage().getCannonDesignList()) {
-			// check of all directions
-			BlockFace cannonDirection = BlockFace.NORTH;
-			for (int i = 0; i < 4; i++) {
-				// for all blocks for the design
-				List<SimpleBlock> designBlockList = cannonDesign.getAllCannonBlocks(cannonDirection);
+        // check all cannon design if this block is part of the design
+        for (CannonDesign cannonDesign : plugin.getDesignStorage().getCannonDesignList()) {
+            // check of all directions
+            BlockFace cannonDirection = BlockFace.NORTH;
+            for (int i = 0; i < 4; i++) {
+                // for all blocks for the design
+                List<SimpleBlock> designBlockList = cannonDesign.getAllCannonBlocks(cannonDirection);
                 //check for empty entries
                 if (designBlockList.isEmpty()) {
                     plugin.logSevere("There are empty cannon design schematics in your design folder. Please check it.");
                     return null;
                 }
 
-				for (SimpleBlock designBlock : designBlockList) {
-					// compare blocks
+                for (SimpleBlock designBlock : designBlockList) {
+                    // compare blocks
                     if (!designBlock.compareMaterialAndFacing(cannonBlock.getBlock().getBlockData())) {
                         continue;
                     }
@@ -625,80 +656,81 @@ public class CannonManager {
                         return new Cannon(cannonDesign, world.getUID(), offset, cannonDirection, owner);
                     }
                 }
-				// rotate cannon direction
-				cannonDirection = CannonsUtil.roatateFace(cannonDirection);
-			}
+                // rotate cannon direction
+                cannonDirection = CannonsUtil.roatateFace(cannonDirection);
+            }
 
-		}
-		return null;
-	}
+        }
+        return null;
+    }
 
-	/**
-	 * returns the number of owned cannons of a player
-	 * @param player the owner of the cannons
-	 * @return number of cannons
-	 */
-	public int getNumberOfCannons(UUID player) {
-		int i = 0;
-		for (Cannon cannon : cannonList.values()) {
-			if (cannon.getOwner() == null) {
-				plugin.logSevere("Cannon has no owner. Contact the plugin developer");
-			} else if (cannon.getOwner().equals(player)) {
-				i++;
-			}
-		}
-		return i;
-	}
+    /**
+     * returns the number of owned cannons of a player
+     *
+     * @param player the owner of the cannons
+     * @return number of cannons
+     */
+    public int getNumberOfCannons(UUID player) {
+        int i = 0;
+        for (Cannon cannon : cannonList.values()) {
+            if (cannon.getOwner() == null) {
+                plugin.logSevere("Cannon has no owner. Contact the plugin developer");
+            } else if (cannon.getOwner().equals(player)) {
+                i++;
+            }
+        }
+        return i;
+    }
 
-	/**
-	 * 
-	 * @return List of cannons
-	 */
-	public static ConcurrentHashMap<UUID,Cannon> getCannonList() {
-		return cannonList;
-	}
-	
-	/**
-	 * List of cannons
-	 */
-	public void clearCannonList() {
-		cannonList.clear();
-	}
+    /**
+     * @return List of cannons
+     */
+    public static ConcurrentHashMap<UUID, Cannon> getCannonList() {
+        return cannonList;
+    }
 
-	/**
-	 * returns the number of cannons manged by the plugin
-	 * @return number of cannons in all world
-	 */
-	public int getCannonListSize() {
-		return cannonList.size();
-	}
+    /**
+     * List of cannons
+     */
+    public void clearCannonList() {
+        cannonList.clear();
+    }
 
-	/**
-	 * returns the amount of cannons a player can build
-	 * 
-	 * @param player check the number of cannon this player can build
-	 * @return the maximum number of cannons
-	 */
+    /**
+     * returns the number of cannons manged by the plugin
+     *
+     * @return number of cannons in all world
+     */
+    public int getCannonListSize() {
+        return cannonList.size();
+    }
+
+    /**
+     * returns the amount of cannons a player can build
+     *
+     * @param player check the number of cannon this player can build
+     * @return the maximum number of cannons
+     */
     public int getCannonBuiltLimit(Player player) {
-		// the player is not valid - no limit check
-		if (player == null) return Integer.MAX_VALUE;
+        // the player is not valid - no limit check
+        if (player == null) return Integer.MAX_VALUE;
 
-		// both limitA/B and cannons.limit.5 work
-		// if all notes are enabled, set limit to a high number. If no permission plugin is loaded, everything is enabled
+        // both limitA/B and cannons.limit.5 work
+        // if all notes are enabled, set limit to a high number. If no permission plugin is loaded, everything is enabled
 
         int newBuiltLimit = getNewBuildLimit(player);
 
-		// config implementation
-		if (config.isBuildLimitEnabled()) {
+        // config implementation
+        if (config.isBuildLimitEnabled()) {
             plugin.logDebug("BuildLimit: limitA and limitB are enabled");
             int buildLimitB = config.getBuildLimitB();
             int buildLimitA = config.getBuildLimitA();
 
-			if (player.hasPermission("cannons.limit.limitB") && (newBuiltLimit > buildLimitB)) {
-				// return the
+            if (player.hasPermission("cannons.limit.limitB") && (newBuiltLimit > buildLimitB)) {
+                // return the
                 plugin.logDebug("build limitB sets the number of cannons to: " + buildLimitB);
-				return buildLimitB;
-			}
+                return buildLimitB;
+            }
 
             if (player.hasPermission("cannons.limit.limitA") && (newBuiltLimit > buildLimitA)) {
                 // limit B is stronger
@@ -706,17 +738,18 @@ public class CannonManager {
                 return buildLimitA;
             }
         }
-		// player implementation
-		if (newBuiltLimit >= 0) {
+        // player implementation
+        if (newBuiltLimit >= 0) {
             plugin.logDebug("permission build limit sets the maximum number of cannons to: " + newBuiltLimit);
             return newBuiltLimit;
-		}
+        }
         plugin.logDebug("no build limit found. Setting to max value.");
         return Integer.MAX_VALUE;
-	}
+    }
 
     /**
      * returns the build limit for the player given by cannons.limit.5
+     *
      * @param player the build limit for this player
      * @return how many cannosn this player can build
      */
@@ -740,28 +773,29 @@ public class CannonManager {
         plugin.logDebug("BuildLimit: no entry for cannons.limit.x found");
         return Integer.MAX_VALUE;
     }
-	/**
-	 * checks if the player can build a cannon, checks permission and builtLimit
-	 */
-	private MessageEnum canBuildCannon(Cannon cannon, UUID owner) {
-		CannonDesign design = cannon.getCannonDesign();
-		
-		//get the player from the server
-		if (owner == null) return null;
-		Player player = Bukkit.getPlayer(owner);
-		if (player == null) return null;
-	
-		// check if player has permission to build
-		if (!player.hasPermission(design.getPermissionBuild())) {
-			return MessageEnum.PermissionErrorBuild;
-		}
-		// player does not have too many guns
-		if (getNumberOfCannons(owner) >= getCannonBuiltLimit(player)) {
-			return MessageEnum.ErrorCannonBuiltLimit;
-		}
-		// player has sufficient permission to build a cannon
-		return MessageEnum.CannonCreated;
-	}
+
+    /**
+     * checks if the player can build a cannon, checks permission and builtLimit
+     */
+    private MessageEnum canBuildCannon(Cannon cannon, UUID owner) {
+        CannonDesign design = cannon.getCannonDesign();
+
+        //get the player from the server
+        if (owner == null) return null;
+        Player player = Bukkit.getPlayer(owner);
+        if (player == null) return null;
+
+        // check if player has permission to build
+        if (!player.hasPermission(design.getPermissionBuild())) {
+            return MessageEnum.PermissionErrorBuild;
+        }
+        // player does not have too many guns
+        if (getNumberOfCannons(owner) >= getCannonBuiltLimit(player)) {
+            return MessageEnum.ErrorCannonBuiltLimit;
+        }
+        // player has sufficient permission to build a cannon
+        return MessageEnum.CannonCreated;
+    }
 
     /**
      * removes all cannon
@@ -781,25 +815,26 @@ public class CannonManager {
     }
 
 
-	/**
-	 * Deletes all cannons of this player in the database to reset the cannon limit
-	 * @param owner the owner of the cannon
+    /**
+     * Deletes all cannons of this player in the database to reset the cannon limit
+     *
+     * @param owner the owner of the cannon
      * @return returns true if there was an entry of this player in the list
-	 */
-	public boolean deleteCannons(UUID owner) {
-		Iterator<Cannon> iter = cannonList.values().iterator();
+     */
+    public boolean deleteCannons(UUID owner) {
+        Iterator<Cannon> iter = cannonList.values().iterator();
         boolean inList = false;
 
-		while (iter.hasNext()) {
-			Cannon next = iter.next();
-			if (next.getOwner() != null && next.getOwner().equals(owner)) {
+        while (iter.hasNext()) {
+            Cannon next = iter.next();
+            if (next.getOwner() != null && next.getOwner().equals(owner)) {
                 inList = true;
-				next.destroyCannon(false, false, BreakCause.Other);
-				iter.remove();
-			}
-		}
+                next.destroyCannon(false, false, BreakCause.Other);
+                iter.remove();
+            }
+        }
         return inList;
-	}
+    }
 
     /**
      * reloads designs from the design list and updates all entries in the cannon
