+++ src/main/java/at/pavlov/cannons/listener/Commands.java
@@ -13,14 +13,13 @@ import at.pavlov.cannons.config.UserMessages;
 import at.pavlov.cannons.dao.PersistenceDatabase;
 import at.pavlov.cannons.projectile.Projectile;
 import at.pavlov.cannons.projectile.ProjectileStorage;
+import at.pavlov.cannons.utils.CannonSelector;
 import at.pavlov.cannons.utils.CannonsUtil;
 import com.google.common.base.Joiner;
-import net.milkbowl.vault.economy.EconomyResponse;
 import org.apache.commons.lang3.StringUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.OfflinePlayer;
-import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -28,10 +27,8 @@ import org.bukkit.command.TabExecutor;
 import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.UUID;
 
 
 public class Commands implements TabExecutor {
@@ -39,20 +36,14 @@ public class Commands implements TabExecutor {
     private final Config config;
     private final UserMessages userMessages;
     private final PersistenceDatabase persistenceDatabase;
-
-    //<player,command to be performed>;
-    private HashMap<UUID, SelectCannon> cannonSelector = new HashMap<>();
-    //<player,command to be performed>;
-    private HashMap<UUID, Boolean> selectTargetBoolean = new HashMap<>();
-    //<player,playerUID>;
-    private HashMap<UUID, UUID> whitelistPlayer = new HashMap<>();
-
+    private final CannonSelector selector;
 
     public Commands(Cannons plugin) {
         this.plugin = plugin;
         config = this.plugin.getMyConfig();
         userMessages = this.plugin.getMyConfig().getUserMessages();
         persistenceDatabase = this.plugin.getPersistenceDatabase();
+        selector = CannonSelector.getInstance();
     }
 
 
@@ -316,7 +307,7 @@ public class Commands implements TabExecutor {
                 plugin.logDebug(tag + sender.getName() + noPerm + args[0]);
                 return true;
             }
-            toggleBuyCannon(player, SelectCannon.BUY_CANNON);
+            selector.toggleBuyCannon(player, SelectCannon.BUY_CANNON);
             return true;
         }
         //rename cannon
@@ -345,7 +336,7 @@ public class Commands implements TabExecutor {
             if (args.length >= 2 && (args[1].equalsIgnoreCase("off") || args[1].equalsIgnoreCase("disable") || args[1].equalsIgnoreCase("remove")))
                 plugin.getAiming().removeObserverForAllCannons(player);
             else if (args.length < 2)
-                toggleCannonSelector(player, SelectCannon.OBSERVER);
+                selector.toggleCannonSelector(player, SelectCannon.OBSERVER);
             else if (args.length >= 2 && args[1] != null) {
                 //selection done by a string '/cannons observer CANNON_NAME'
                 Cannon cannon = CannonManager.getCannon(args[1]);
@@ -366,15 +357,13 @@ public class Commands implements TabExecutor {
             if (args.length >= 3 && (args[1].equalsIgnoreCase("add"))) {
                 OfflinePlayer offPlayer = CannonsUtil.getOfflinePlayer(args[2]);
                 if (offPlayer != null && offPlayer.hasPlayedBefore()) {
-                    whitelistPlayer.put(player.getUniqueId(), offPlayer.getUniqueId());
-                    toggleCannonSelector(player, SelectCannon.WHITELIST_ADD);
+                    selector.toggleCannonSelector(player, SelectCannon.WHITELIST_ADD, offPlayer);
                 } else
                     userMessages.sendMessage(MessageEnum.ErrorPlayerNotFound, player);
             } else if (args.length >= 3 && (args[1].equalsIgnoreCase("remove"))) {
                 OfflinePlayer offPlayer = CannonsUtil.getOfflinePlayer(args[2]);
                 if (offPlayer != null && offPlayer.hasPlayedBefore()) {
-                    whitelistPlayer.put(player.getUniqueId(), offPlayer.getUniqueId());
-                    toggleCannonSelector(player, SelectCannon.WHITELIST_REMOVE);
+                    selector.toggleCannonSelector(player, SelectCannon.WHITELIST_REMOVE, offPlayer);
                 } else
                     userMessages.sendMessage(MessageEnum.ErrorPlayerNotFound, player);
             } else
@@ -390,7 +379,7 @@ public class Commands implements TabExecutor {
             boolean choice = false;
             if (args.length >= 3) {
                 choice = Boolean.parseBoolean(args[2]);
-                selectTargetBoolean.put(player.getUniqueId(), choice);
+                selector.putTarget(player.getUniqueId(), choice);
             }
             // additional range command to select multiple cannons
             int length = 0;
@@ -406,25 +395,25 @@ public class Commands implements TabExecutor {
                 if (length > 0)
                     selectCannonsInBox(player, SelectCannon.TARGET_MOB, length);
                 else
-                    toggleCannonSelector(player, SelectCannon.TARGET_MOB);
+                    selector.toggleCannonSelector(player, SelectCannon.TARGET_MOB);
             } else if (args.length >= 2 && (args[1].equalsIgnoreCase("player"))) {
                 if (length > 0)
                     selectCannonsInBox(player, SelectCannon.TARGET_PLAYER, length);
                 else
-                    toggleCannonSelector(player, SelectCannon.TARGET_PLAYER);
+                    selector.toggleCannonSelector(player, SelectCannon.TARGET_PLAYER);
             } else if (args.length >= 2 && (args[1].equalsIgnoreCase("cannon"))) {
                 if (length > 0)
                     selectCannonsInBox(player, SelectCannon.TARGET_CANNON, length);
                 else
-                    toggleCannonSelector(player, SelectCannon.TARGET_CANNON);
+                    selector.toggleCannonSelector(player, SelectCannon.TARGET_CANNON);
             } else if (args.length >= 2 && (args[1].equalsIgnoreCase("other"))) {
                 if (length > 0)
                     selectCannonsInBox(player, SelectCannon.TARGET_OTHER, length);
                 else
-                    toggleCannonSelector(player, SelectCannon.TARGET_OTHER);
+                    selector.toggleCannonSelector(player, SelectCannon.TARGET_OTHER);
             } else {
                 //remove choice for target it the command was invalid
-                selectTargetBoolean.remove(player.getUniqueId());
+                selector.removeTarget(player.getUniqueId());
                 sendMessage(sender, ChatColor.RED + "Usage '/cannons target <mob|player|cannon|other> <true|false> <range>'");
             }
         }
@@ -434,7 +423,7 @@ public class Commands implements TabExecutor {
                 plugin.logDebug(tag + sender.getName() + noPerm + args[0]);
                 return true;
             }
-            toggleCannonSelector(player, SelectCannon.INFO);
+            selector.toggleCannonSelector(player, SelectCannon.INFO);
         }
         //get name of cannon
         else if (args[0].equalsIgnoreCase("dismantle")) {
@@ -442,7 +431,7 @@ public class Commands implements TabExecutor {
                 plugin.logDebug(tag + sender.getName() + noPerm + args[0]);
                 return true;
             }
-            toggleCannonSelector(player, SelectCannon.DISMANTLE);
+            selector.toggleCannonSelector(player, SelectCannon.DISMANTLE);
         }
         //list cannons of this player name
         else if (args[0].equalsIgnoreCase("list")) {
@@ -484,7 +473,7 @@ public class Commands implements TabExecutor {
                 plugin.logDebug(tag + sender.getName() + noPerm + args[0]);
                 return true;
             }
-            toggleCannonSelector(player, SelectCannon.BLOCK_DATA);
+            selector.toggleCannonSelector(player, SelectCannon.BLOCK_DATA);
         }
         //claim cannons in the surrounding
         else if (args[0].equalsIgnoreCase(CommandList.CLAIM.getCommand())) {
@@ -526,325 +515,6 @@ public class Commands implements TabExecutor {
     }
 
     /**
-     * this player will be removed from the selecting mode
-     *
-     * @param player the player will be removed
-     * @param cmd    this command will be performed when the cannon is selected
-     */
-    public void addCannonSelector(Player player, SelectCannon cmd) {
-        if (player == null || cmd == null)
-            return;
-
-        if (!isSelectingMode(player)) {
-            cannonSelector.put(player.getUniqueId(), cmd);
-            if (isBlockSelectingMode(player))
-                userMessages.sendMessage(MessageEnum.CmdSelectBlock, player);
-            else
-                userMessages.sendMessage(MessageEnum.CmdSelectCannon, player);
-        }
-    }
-
-    /**
-     * this player will be removed from the selecting mode
-     *
-     * @param player the player will be removed
-     */
-    public void removeCannonSelector(Player player) {
-        if (player == null)
-            return;
-
-        if (isSelectingMode(player)) {
-            cannonSelector.remove(player.getUniqueId());
-            userMessages.sendMessage(MessageEnum.CmdSelectCanceled, player);
-        }
-    }
-
-    /**
-     * selecting mode will be toggled
-     *
-     * @param player the player using the selecting mode
-     * @param cmd    this command will be performed when the cannon is selected
-     */
-    public void toggleCannonSelector(Player player, SelectCannon cmd) {
-        if (player == null)
-            return;
-
-        if (isSelectingMode(player))
-            removeCannonSelector(player);
-        else
-            addCannonSelector(player, cmd);
-    }
-
-    /**
-     * this player will be removed from the buying mode
-     *
-     * @param player the player will be removed
-     * @param cmd    this command will be performed when the cannon is selected
-     */
-    public void addBuyCannon(Player player, SelectCannon cmd) {
-        if (player == null || cmd == null)
-            return;
-
-        if (!isSelectingMode(player)) {
-            cannonSelector.put(player.getUniqueId(), cmd);
-            userMessages.sendMessage(MessageEnum.CmdBuyCannon, player);
-        }
-    }
-
-    /**
-     * this player will be removed from the buying mode
-     *
-     * @param player the player will be removed
-     */
-    public void removeBuyCannon(Player player) {
-        if (player == null)
-            return;
-
-        if (isSelectingMode(player)) {
-            cannonSelector.remove(player.getUniqueId());
-            userMessages.sendMessage(MessageEnum.CmdSelectCanceled, player);
-        }
-    }
-
-    /**
-     * buying mode will be toggled
-     *
-     * @param player the player using the selecting mode
-     * @param cmd    this command will be performed when the cannon is selected
-     */
-    public void toggleBuyCannon(Player player, SelectCannon cmd) {
-        if (player == null)
-            return;
-
-        if (isSelectingMode(player))
-            removeBuyCannon(player);
-        else
-            addBuyCannon(player, cmd);
-    }
-
-
-    /**
-     * Checks if this player is in selecting mode
-     *
-     * @param player player to check
-     * @return true if in selecting mode
-     */
-    public boolean isSelectingMode(Player player) {
-        return player != null && cannonSelector.containsKey(player.getUniqueId());
-    }
-
-    public boolean isBlockSelectingMode(Player player) {
-        SelectCannon cmd = cannonSelector.get(player.getUniqueId());
-        return cmd.equals(SelectCannon.BLOCK_DATA);
-    }
-
-    /**
-     * adds a new selected cannon for this player
-     *
-     * @param player player that selected the cannon
-     * @param block  the selected block
-     */
-    public void setSelectedBlock(Player player, Block block) {
-        if (player == null || block == null)
-            return;
-
-        SelectCannon cmd = cannonSelector.get(player.getUniqueId());
-        if (cmd != null) {
-            switch (cmd) {
-                case BLOCK_DATA: {
-                    player.sendMessage(block.getBlockData().getAsString());
-                    break;
-                }
-            }
-        }
-        cannonSelector.remove(player.getUniqueId());
-    }
-
-
-    /**
-     * adds a new selected cannon for this player
-     *
-     * @param player player that selected the cannon
-     * @param cannon the selected cannon
-     */
-    public void setSelectedCannon(Player player, Cannon cannon) {
-        if (player == null || cannon == null)
-            return;
-
-        SelectCannon cmd = cannonSelector.get(player.getUniqueId());
-
-        if (cmd == null) {
-            selectTargetBoolean.remove(player.getUniqueId());
-            cannonSelector.remove(player.getUniqueId());
-            return;
-        }
-
-        switch (cmd) {
-            case OBSERVER: {
-                MessageEnum message = cannon.toggleObserver(player, false);
-                userMessages.sendMessage(message, player, cannon);
-                CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-                break;
-            }
-            case INFO: {
-                userMessages.sendMessage(MessageEnum.CannonInfo, player, cannon);
-                CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-                break;
-            }
-            case DISMANTLE: {
-                plugin.getCannonManager().dismantleCannon(cannon, player);
-                break;
-            }
-            case WHITELIST_ADD: {
-                if (!cannon.getCannonDesign().isSentry()) {
-                    userMessages.sendMessage(MessageEnum.CmdNoSentryWhitelist, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-
-                if (!player.getUniqueId().equals(cannon.getOwner())) {
-                    userMessages.sendMessage(MessageEnum.ErrorNotTheOwner, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                } else {
-                    cannon.addWhitelistPlayer(whitelistPlayer.get(player.getUniqueId()));
-                    whitelistPlayer.remove(player.getUniqueId());
-                    userMessages.sendMessage(MessageEnum.CmdAddedWhitelist, player, cannon);
-                    CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-                }
-                break;
-            }
-            case WHITELIST_REMOVE: {
-                if (!cannon.getCannonDesign().isSentry()) {
-                    userMessages.sendMessage(MessageEnum.CmdNoSentryWhitelist, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-
-                if (!player.getUniqueId().equals(cannon.getOwner())) {
-                    userMessages.sendMessage(MessageEnum.ErrorNotTheOwner, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-
-                if (player.getUniqueId() != cannon.getOwner()) {
-                    cannon.removeWhitelistPlayer(whitelistPlayer.get(player.getUniqueId()));
-                    whitelistPlayer.remove(player.getUniqueId());
-                    userMessages.sendMessage(MessageEnum.CmdRemovedWhitelist, player, cannon);
-                } else {
-                    userMessages.sendMessage(MessageEnum.CmdRemovedWhitelistOwner, player, cannon);
-                }
-                CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-                break;
-            }
-            case TARGET_MOB: {
-                if (!player.getUniqueId().equals(cannon.getOwner())) {
-                    userMessages.sendMessage(MessageEnum.ErrorNotTheOwner, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-
-                if (!cannon.getCannonDesign().isSentry()) {
-                    break;
-                }
-                // use preselected choice or toggle
-                if (selectTargetBoolean.containsKey(player.getUniqueId()))
-                    cannon.setTargetMob(selectTargetBoolean.get(player.getUniqueId()));
-                else
-                    cannon.toggleTargetMob();
-                userMessages.sendMessage(MessageEnum.CmdToggledTargetMob, player, cannon);
-                CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-
-                break;
-            }
-            case TARGET_PLAYER: {
-                if (!player.getUniqueId().equals(cannon.getOwner())) {
-                    userMessages.sendMessage(MessageEnum.ErrorNotTheOwner, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-
-                if (!cannon.getCannonDesign().isSentry()) {
-                    break;
-                }
-                // use preselected choice or toggle
-                if (selectTargetBoolean.containsKey(player.getUniqueId()))
-                    cannon.setTargetPlayer(selectTargetBoolean.get(player.getUniqueId()));
-                else
-                    cannon.toggleTargetPlayer();
-                userMessages.sendMessage(MessageEnum.CmdToggledTargetPlayer, player, cannon);
-                CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-
-                break;
-            }
-            case TARGET_CANNON: {
-                if (cannon.getCannonDesign().isSentry() && !player.getUniqueId().equals(cannon.getOwner())) {
-                    userMessages.sendMessage(MessageEnum.ErrorNotTheOwner, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-                if (!cannon.getCannonDesign().isSentry()) {
-                    break;
-                }
-                // use preselected choice or toggle
-                if (selectTargetBoolean.containsKey(player.getUniqueId()))
-                    cannon.setTargetCannon(selectTargetBoolean.get(player.getUniqueId()));
-                else
-                    cannon.toggleTargetCannon();
-                userMessages.sendMessage(MessageEnum.CmdToggledTargetCannon, player, cannon);
-                CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-
-                break;
-            }
-            case TARGET_OTHER: {
-                if (cannon.getCannonDesign().isSentry() && !player.getUniqueId().equals(cannon.getOwner())) {
-                    userMessages.sendMessage(MessageEnum.ErrorNotTheOwner, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-
-                if (!cannon.getCannonDesign().isSentry()) {
-                    break;
-                }
-                // use preselected choice or toggle
-                if (selectTargetBoolean.containsKey(player.getUniqueId()))
-                    cannon.setTargetOther(selectTargetBoolean.get(player.getUniqueId()));
-                else
-                    cannon.toggleTargetOther();
-                userMessages.sendMessage(MessageEnum.CmdToggledTargetOther, player, cannon);
-                CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-
-                break;
-            }
-            case BUY_CANNON: {
-                if (cannon.isPaid()) {
-                    userMessages.sendMessage(MessageEnum.ErrorAlreadyPaid, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                    break;
-                }
-                //redraw money if required
-                if (plugin.getEconomy() == null || !(cannon.getCannonDesign().getEconomyBuildingCost() > 0)) {
-                    break;
-                }
-
-                EconomyResponse r = plugin.getEconomy().withdrawPlayer(player, cannon.getCannonDesign().getEconomyBuildingCost());
-
-                if (!r.transactionSuccess()) {
-                    userMessages.sendMessage(MessageEnum.ErrorNoMoney, player, cannon);
-                    CannonsUtil.playErrorSound(cannon.getMuzzle());
-                } else {
-                    cannon.boughtByPlayer(player.getUniqueId());
-                    //CannonsUtil.playSound();
-                    userMessages.sendMessage(MessageEnum.CmdPaidCannon, player, cannon);
-                    CannonsUtil.playSound(cannon.getMuzzle(), cannon.getCannonDesign().getSoundSelected());
-                }
-                break;
-            }
-        }
-        selectTargetBoolean.remove(player.getUniqueId());
-        cannonSelector.remove(player.getUniqueId());
-    }
-
-    /**
      * @param player player for selecting the cannon
      * @param cmd    select command to perform
      * @param length edge length of of the box
@@ -858,14 +528,12 @@ public class Commands implements TabExecutor {
 
         boolean choice = true;
         //buffer the selection because it will be reset after every cannon
-        if (selectTargetBoolean.containsKey(player.getUniqueId()))
-            choice = selectTargetBoolean.get(player.getUniqueId());
+        if (selector.containsTarget(player.getUniqueId()))
+            choice = selector.getTarget(player.getUniqueId());
 
         HashSet<Cannon> list = CannonManager.getCannonsInBox(player.getLocation(), length, length, length);
         for (Cannon cannon : list) {
-            selectTargetBoolean.put(player.getUniqueId(), choice);
-            cannonSelector.put(player.getUniqueId(), cmd);
-            setSelectedCannon(player, cannon);
+            selector.setSelectedCannon(player, cannon, cmd, choice);
         }
     }
